requirements.txt
Django>=3.2.9
ipython>=7.29.0
python-dotenv>=0.19.2
djangorestframework>=3.12.4
Markdown>=3.3.6
django-filter>=21.1
